+++ design-components/packages/railz-components-angular/src/lib/stencil-generated/index.ts
@@ -12,5 +12,6 @@ export const DIRECTIVES = [
   d.RailzInputText,
   d.RailzInputToggle,
   d.RailzLogo,
+  d.RailzNavigationProgressBar,
   d.RailzStatusPill
 ];